made fixes as requested
