+++ crates/node/Cargo.toml
@@ -25,7 +25,6 @@ webbrowser.workspace = true
 calimero-context.workspace = true
 calimero-context-config.workspace = true
 calimero-crypto.workspace = true
-calimero-store-rocksdb.workspace = true
 calimero-blobstore.workspace = true
 calimero-network.workspace = true
 calimero-node-primitives.workspace = true
@@ -33,6 +32,7 @@ calimero-primitives = { workspace = true, features = ["borsh"] }
 calimero-runtime.workspace = true
 calimero-server = { workspace = true, features = ["jsonrpc", "websocket", "admin"] }
 calimero-store = { workspace = true, features = ["datatypes"] }
+calimero-store-rocksdb.workspace = true
 
 [lints]
 workspace = true
